+++ doc-src/auto_examples/plot_figure.py
@@ -64,7 +64,7 @@ with slide('A bokeh figure'):
     y = np.random.rand(100)
     p.circle(x, y, legend="sin(x)")
 
-    # figure(p)
+    figure(p)
 
 # Export the 3 slides of the presentation
 save('./examples_html_outputs/figure.html')